+++ priv/repo/migrations/20210921153715_add_domain_id_to_grant_requests.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AddDomainIdToGrantRequests do
+  use Ecto.Migration
+
+  def change do
+    alter table("grant_requests") do
+      add(:domain_id, :integer)
+    end
+  end
+end