+++ priv/repo/migrations/20210923205655_remove_grant_requests_updated_at.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.RemoveGrantRequestsUpdatedAt do
+  use Ecto.Migration
+
+  def change do
+    alter table("grant_requests") do
+      remove(:updated_at, :utc_datetime_usec)
+    end
+  end
+end