+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-4076] Support for grant request approval
+- [TD-4113] Jobs change status is included in Audit Events
+
 ### Changed
 
 - [TD-4079] Give grant permission only if we have `gr` templates